Extract dialog component from InspectionItemDialog
